+++ src/cc/lib/graph/metadata.cc
@@ -7,8 +7,12 @@
 #include <cstdio>
 #include <cstring>
 #include <filesystem>
+#include <fstream>
 #include <glog/logging.h>
 #include <glog/raw_logging.h>
+#include <nlohmann/json.hpp>
+
+using json = nlohmann::json;
 
 namespace snark
 {
@@ -21,7 +25,7 @@ Metadata::Metadata(std::filesystem::path path, std::string config_path)
         RAW_LOG_FATAL("HDFS streaming only supported on Linux!");
 #else
     {
-        auto full_path = path / "meta.txt";
+        auto full_path = path / "meta.json";
         auto buffer = read_hdfs<char>(full_path.string(), m_config_path);
 
         path = std::filesystem::temp_directory_path();
@@ -37,14 +41,12 @@ Metadata::Metadata(std::filesystem::path path, std::string config_path)
     }
 #endif
 
-    auto meta = open_meta(std::move(path), "rt");
-    if (fscanf(meta, "v%zu\n", &m_version) <= 0)
-    {
-        RAW_LOG_ERROR(
-            "Failed to read binary data version from meta file. Please use latest deepgnn package to convert data.");
-        exit(errno);
-    }
+    std::ifstream f(path / "meta.json");
+    json meta = json::parse(f);
 
+    std::string version_full = meta["binary_data_version"];
+    version_full.erase(0, 1);
+    m_version = stoi(version_full);
     if (m_version < MINIMUM_SUPPORTED_VERSION)
     {
         RAW_LOG_FATAL("Unsupported version of binary data %zu. Minimum supported version is %zu. Please use latest "
@@ -52,161 +54,73 @@ Metadata::Metadata(std::filesystem::path path, std::string config_path)
                       m_version, MINIMUM_SUPPORTED_VERSION);
     }
 
-    if (fscanf(meta, "%zu\n", &m_node_count) <= 0)
-    {
-        exit(errno);
-    }
-    if (fscanf(meta, "%zu\n", &m_edge_count) <= 0)
-    {
-        exit(errno);
-    }
-    if (fscanf(meta, "%zu\n", &m_node_type_count) <= 0)
-    {
-        exit(errno);
-    }
-    if (fscanf(meta, "%zu\n", &m_edge_type_count) <= 0)
-    {
-        exit(errno);
-    }
-    if (fscanf(meta, "%zu\n", &m_node_feature_count) <= 0)
-    {
-        exit(errno);
-    }
-    if (fscanf(meta, "%zu\n", &m_edge_feature_count) <= 0)
-    {
-        exit(errno);
-    }
+    m_node_count = meta["node_count"];
+    m_edge_count = meta["edge_count"];
+    m_node_type_count = meta["node_type_count"];
+    m_edge_type_count = meta["edge_type_count"];
+    m_node_feature_count = meta["node_feature_count"];
+    m_edge_feature_count = meta["edge_feature_count"];
 
-    if (fscanf(meta, "%zu\n", &m_partition_count) <= 0)
-    {
-        exit(errno);
-    }
+    m_partition_count = meta["partitions"].size();
 
     m_partition_node_weights =
         std::vector<std::vector<float>>(m_partition_count, std::vector<float>(m_node_type_count, 0.0f));
     m_partition_edge_weights =
         std::vector<std::vector<float>>(m_partition_count, std::vector<float>(m_edge_type_count, 0.0f));
 
-    float edge_weight, node_weight;
-    uint32_t partition_num;
     for (size_t p = 0; p < m_partition_count; ++p)
     {
-        if (fscanf(meta, "%ul\n", &partition_num) <= 0)
-        {
-            exit(errno);
-        }
         for (size_t i = 0; i < m_node_type_count; ++i)
         {
-            if (fscanf(meta, "%f\n", &node_weight) <= 0)
-            {
-                exit(errno);
-            }
-            m_partition_node_weights[partition_num][i] = node_weight;
+            m_partition_node_weights[p][i] = meta["partitions"][std::to_string(p)]["node_weight"][i];
         }
         for (size_t i = 0; i < m_edge_type_count; ++i)
         {
-            if (fscanf(meta, "%f\n", &edge_weight) <= 0)
-            {
-                exit(errno);
-            }
-            m_partition_edge_weights[partition_num][i] = edge_weight;
+            m_partition_edge_weights[p][i] = meta["partitions"][std::to_string(p)]["edge_weight"][i];
         }
     }
-    size_t count;
+
     m_node_count_per_type.resize(m_node_type_count);
     for (size_t i = 0; i < m_node_type_count; ++i)
     {
-        if (fscanf(meta, "%zu\n", &count) <= 0)
-        {
-            exit(errno);
-        }
-        m_node_count_per_type[i] = count;
+        m_node_count_per_type[i] = meta["node_count_per_type"][i];
     }
     m_edge_count_per_type.resize(m_edge_type_count);
     for (size_t i = 0; i < m_edge_type_count; ++i)
     {
-        if (fscanf(meta, "%zu\n", &count) <= 0)
-        {
-            exit(errno);
-        }
-        m_edge_count_per_type[i] = count;
+        m_edge_count_per_type[i] = meta["edge_count_per_type"][i];
     }
-    fclose(meta);
 }
 
 void Metadata::Write(std::filesystem::path path) const
 {
-    auto meta = open_meta(std::move(path), "w+");
-    if (fprintf(meta, "v%zu\n", m_version) <= 0)
-    {
-        exit(errno);
-    }
-    if (fprintf(meta, "%zu\n", m_node_count) <= 0)
-    {
-        exit(errno);
-    }
-    if (fprintf(meta, "%zu\n", m_edge_count) <= 0)
-    {
-        exit(errno);
-    }
-    if (fprintf(meta, "%zu\n", m_node_type_count) <= 0)
-    {
-        exit(errno);
-    }
-    if (fprintf(meta, "%zu\n", m_edge_type_count) <= 0)
-    {
-        exit(errno);
-    }
-    if (fprintf(meta, "%zu\n", m_node_feature_count) <= 0)
-    {
-        exit(errno);
-    }
-    if (fprintf(meta, "%zu\n", m_edge_feature_count) <= 0)
-    {
-        exit(errno);
-    }
+    std::string version_str = "v";
+    version_str += std::to_string(m_version);
 
-    if (fprintf(meta, "%zu\n", m_partition_count) <= 0)
-    {
-        exit(errno);
-    }
+    json json_meta = {
+        {"binary_data_version", version_str},
+        {"node_count", m_node_count},
+        {"edge_count", m_edge_count},
+        {"node_type_count", m_node_type_count},
+        {"edge_type_count", m_edge_type_count},
+        {"node_feature_count", m_node_feature_count},
+        {"edge_feature_count", m_edge_feature_count},
+    };
 
+    json_meta["partitions"] = {{"0", {{"node_weight", {0}}}}};
     for (size_t p = 0; p < m_partition_count; ++p)
     {
-        if (fprintf(meta, "%zu\n", p) <= 0)
-        {
-            exit(errno);
-        }
-        for (size_t i = 0; i < m_node_type_count; ++i)
-        {
-            if (fprintf(meta, "%f\n", m_partition_node_weights[p][i]) <= 0)
-            {
-                exit(errno);
-            }
-        }
-        for (size_t i = 0; i < m_edge_type_count; ++i)
-        {
-            if (fprintf(meta, "%f\n", m_partition_edge_weights[p][i]) <= 0)
-            {
-                exit(errno);
-            }
-        }
+        json_meta["partitions"][std::to_string(p)] = {{"node_weight", {0}}};
+        json_meta["partitions"][std::to_string(p)]["node_weight"] = m_partition_node_weights[p];
+        json_meta["partitions"][std::to_string(p)]["edge_weight"] = m_partition_edge_weights[p];
     }
-    for (size_t i = 0; i < m_node_type_count; ++i)
-    {
-        if (fprintf(meta, "%zu\n", m_node_count_per_type[i]) <= 0)
-        {
-            exit(errno);
-        }
-    }
-    for (size_t i = 0; i < m_edge_type_count; ++i)
-    {
-        if (fprintf(meta, "%zu\n", m_edge_count_per_type[i]) <= 0)
-        {
-            exit(errno);
-        }
-    }
-    fclose(meta);
+
+    json_meta["node_count_per_type"] = m_node_count_per_type;
+    json_meta["edge_count_per_type"] = m_edge_count_per_type;
+
+    std::ofstream meta(path / "meta.json");
+    meta << json_meta;
+    meta.close();
 }
 
 } // namespace snark